refactor: MenuRequestDto 위치 변경
